fix link error when not defining VSTGUI_DIRECT2D_SUPPORT

+++ vstgui/lib/platform/win32/win32support.h
@@ -166,7 +166,7 @@ protected:
 	HGLOBAL resData;
 	uint32_t streamPos;
 	uint32_t resSize;
-	uint32_t _refcount;
+	LONG _refcount;
 };
 
 /// @endcond
